+++ src/toolchain/source/mod.rs
@@ -6,7 +6,7 @@ pub use source_buffer::SourceBuffer;
 macro_rules! sclang {
     ( $s:literal ) => {
         &source::SourceBuffer::new_from_string(
-            $s,
+            bstr::BStr::new($s),
             const_format::formatcp!("{}:{}:{}", file!(), line!(), column!()),
         )
         .unwrap()